requirements.txt
asgiref==3.8.1
channels==4.0.0
Django==5.0.3
djangorestframework==3.15.1
sqlparse==0.4.4
typing_extensions==4.10.0
